+++ R/generate_data.R
@@ -4,7 +4,9 @@
 #' @param dtDefs name of definitions data.table/data.frame. If no definitions
 #'  are provided
 #' a data set with ids only is generated.
-#' @param id The string defining the id of the record
+#' @param id The string defining the id of the record. Will override previously
+#'  set id name with a warning (unless the old value is 'id'). If the 
+#' id attribute in dtDefs is NULL will default to 'id'.
 #' @param envir Environment the data definitions are evaluated in.
 #'  Defaults to [base::parent.frame].
 #' @return A data.table that contains the simulated data.
@@ -59,10 +61,25 @@ genData <- function(n, dtDefs = NULL, id = "id", envir = parent.frame()) {
   } else { # existing definitions
     assertClass(dtDefs = dtDefs, class = "data.table")
 
-    idname <- attr(dtDefs, "id")
+    oldId <- attr(dtDefs, "id")
+    if (!is.null(oldId) && id != oldId && !missing(id)) {
+      if (oldId != "id") {
+        valueWarning(
+          var = oldId,
+          names = id,
+          msg = list(
+            "Previously defined 'id'-column found: '{var}'. ",
+            "The current specification '{names}' will override it."
+          )
+        )
+      }
+    } else {
+      id <- oldId %||% id
+    }
 
-    dfSimulate <- data.frame(c(1:n)) # initialize simulated data with ids
-    names(dfSimulate) <- attr(dtDefs, "id") # name first column attribute "id"
+    dfSimulate <- data.table::data.table(x = 1:n)
+    data.table::setnames(dfSimulate, id)
+    data.table::setkeyv(dfSimulate, id)
     iter <- nrow(dtDefs) # generate a column of data for each row of dtDefs
 
     for (i in (1:iter)) {
@@ -70,13 +87,13 @@ genData <- function(n, dtDefs = NULL, id = "id", envir = parent.frame()) {
         args = dtDefs[i, ],
         n = n,
         dfSim = dfSimulate,
-        idname = idname,
+        idname = id,
         envir = envir
       )
     }
 
     dt <- data.table::data.table(dfSimulate)
-    data.table::setkeyv(dt, idname)
+    data.table::setkeyv(dt, id)
   }
 
   return(dt[])
@@ -613,7 +630,7 @@ genOrdCat <- function(dtName,
   if (!is.null(adjVar)) {
     adjVar <- ensureLength(
       adjVar = adjVar,
-      n = seq_len(nCats), msg = list(
+      n = nCats, msg = list(
         "Number of categories implied",
         " by baseprobs and adjVar do not match. ",
         "{ dots$names[[1]] } should be",